chore: organize the translation service in a package
